経路クラス変更
